pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "title-image-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[project.scripts]
title-image-service = "title_image_service.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/title_image_service"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
